delete(@vx/boxplot): more boxplot => statsplot
