+++ packages/vx-demo/src/components/gallery.js
@@ -26,7 +26,7 @@ import Trees from './tiles/tree';
 import Cluster from './tiles/dendrogram';
 import Voronoi from './tiles/voronoi';
 import Legends from './tiles/legends';
-import StatsPlots from './tiles/statsplots';
+import StatsPlot from './tiles/statsplot';
 import GeoMercator from './tiles/geo-mercator';
 import GeoCustom from './tiles/geo-custom';
 import Network from './tiles/network';
@@ -507,12 +507,12 @@ export default function() {
         </Tilt>
 
         <Tilt className="tilt" options={{ max: 8, scale: 1 }}>
-          <Link href="/statsplots">
+          <Link href="/statsplot">
             <div className="gallery-item" style={{ background: '#8a88e3' }}>
               <div className="image">
                 <ParentSize>
                   {({ width, height }) => (
-                    <StatsPlots width={width} height={height + detailsHeight} />
+                    <StatsPlot width={width} height={height + detailsHeight} />
                   )}
                 </ParentSize>
               </div>
